chore: pregenerate ref navs
